+++ CopilotKit/packages/react-core/src/index.tsx
@@ -3,3 +3,4 @@ export * from "./context";
 export * from "./hooks";
 export * from "./types";
 export * from "./openai-assistants";
+export * from "./openai";