+++ CopilotKit/examples/next-openai/CHANGELOG.md
@@ -1,5 +1,108 @@
 # web
 
+## 1.13.0
+
+### Minor Changes
+
+- factor useChat into internal core
+- a7b417a: insertion default prompt update
+- 88d6654: release useChat fixes
+- 51de9d5: textarea editing: default prompt + few shot update
+- fa84257: remove vercel ai
+- 98a37c8: strictly propagate copilot api params through the fetch arguments - not through any constructors
+- 250032d: useChat: do not separately propagate options.url to constructor
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [a7b417a]
+- Updated dependencies [88d6654]
+- Updated dependencies [51de9d5]
+- Updated dependencies [fa84257]
+- Updated dependencies [98a37c8]
+- Updated dependencies [250032d]
+  - @copilotkit/react-textarea@0.27.0
+  - @copilotkit/react-core@0.17.0
+  - @copilotkit/react-ui@0.14.0
+
+## 1.13.0-alpha.5
+
+### Minor Changes
+
+- release useChat fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.27.0-alpha.5
+  - @copilotkit/react-core@0.17.0-alpha.5
+  - @copilotkit/react-ui@0.14.0-alpha.5
+
+## 1.13.0-alpha.4
+
+### Minor Changes
+
+- insertion default prompt update
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.27.0-alpha.4
+  - @copilotkit/react-core@0.17.0-alpha.4
+  - @copilotkit/react-ui@0.14.0-alpha.4
+
+## 1.13.0-alpha.3
+
+### Minor Changes
+
+- textarea editing: default prompt + few shot update
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.27.0-alpha.3
+  - @copilotkit/react-core@0.17.0-alpha.3
+  - @copilotkit/react-ui@0.14.0-alpha.3
+
+## 1.13.0-alpha.2
+
+### Minor Changes
+
+- useChat: do not separately propagate options.url to constructor
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.27.0-alpha.2
+  - @copilotkit/react-core@0.17.0-alpha.2
+  - @copilotkit/react-ui@0.14.0-alpha.2
+
+## 1.13.0-alpha.1
+
+### Minor Changes
+
+- strictly propagate copilot api params through the fetch arguments - not through any constructors
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.27.0-alpha.1
+  - @copilotkit/react-core@0.17.0-alpha.1
+  - @copilotkit/react-ui@0.14.0-alpha.1
+
+## 1.13.0-alpha.0
+
+### Minor Changes
+
+- remove vercel ai
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-textarea@0.27.0-alpha.0
+  - @copilotkit/react-core@0.17.0-alpha.0
+  - @copilotkit/react-ui@0.14.0-alpha.0
+
 ## 1.12.0
 
 ### Minor Changes